Refactored plot generation code
