+++ code/model2faster.py
@@ -4,8 +4,9 @@ import torch
 import torch.autograd as autograd
 import torch.nn as nn
 import torch.optim as optim
-import constants
 
+import constants
+import plots
 import utils
 
 from model import Model
@@ -110,7 +111,7 @@ class Model2Faster(nn.Module, Model):
             self.save(file_name)
         email_ids, embs = self.extract_user_embeddings()
         utils.save_user_embeddings(email_ids, embs)
-        utils.plot_with_tsne(email_ids, embs, display_hover=False)
+        plots.plot_with_tsne(email_ids, embs, display_hover=False)
 
     def predict(self, email, w2v):
         loss_criteria = nn.MSELoss()
