+++ code/model4.py
@@ -1,13 +1,16 @@
-from model import Model
+import numpy as np
+import time
 import torch
 import torch.nn as nn
 import torch.autograd as autograd
 import torch.optim as optim
-import utils
-import numpy as np
-import time
-import constants
 import random
+
+import constants
+import plots
+import utils
+
+from model import Model
 
 
 class Model4(nn.Module, Model):
@@ -89,7 +92,7 @@ class Model4(nn.Module, Model):
         email_ids, embs = self.extract_user_embeddings()
         utils.save_user_embeddings(email_ids, embs)
         # utils.get_similar_users(email_ids, embs)
-        utils.plot_with_tsne(email_ids, embs, display_hover=False)
+        plots.plot_with_tsne(email_ids, embs, display_hover=False)
 
 
         # for recv in recv_list:
